!26 Allow for better parsing using StringOrVecToVec
This brings an option to skip empty results when splitting a string
using a separator. This behaviour could occur due to the way the
str::split is implemented in Rust. To prevent this from happening,
this introduces a backwards-compatible work-around which brings another
setting to the StringOrVecToVec structure which is whether to remove
empty elements or not after the string splitting.